Rename tools to services
