+++ backend/src/projects/dydx.ts
@@ -1,4 +1,4 @@
-import { Config, queryTVL } from '../tools'
+import { Config, queryTVL } from '../services'
 
 const BRIDGE_ADDRESS = '0xD54f502e184B6B739d7D27a6410a67dc462D69c8'
