+++ backend/src/projects/fuel.ts
@@ -1,4 +1,4 @@
-import { Config, queryTVL } from '../tools'
+import { Config, queryTVL } from '../services'
 
 const BRIDGE_ADDRESS = '0x6880f6Fd960D1581C2730a451A22EED1081cfD72'
