+++ backend/src/projects/immutablex.ts
@@ -1,4 +1,4 @@
-import { Config, queryTVL } from '../tools'
+import { Config, queryTVL } from '../services'
 
 const BRIDGE_ADDRESS = '0x5FDCCA53617f4d2b9134B29090C87D01058e27e9'
