+++ backend/src/projects/optimism.ts
@@ -1,4 +1,4 @@
-import { Config, queryTVL } from '../tools'
+import { Config, queryTVL } from '../services'
 
 const BRIDGE_ADDRESS_1 = '0x045e507925d2e05D114534D0810a1abD94aca8d6'
 const BRIDGE_ADDRESS_2 = '0x674bdf20A0F284D710BC40872100128e2d66Bd3f'
