+++ backend/src/projects/zksync.ts
@@ -1,4 +1,4 @@
-import { Config, queryTVL } from '../tools'
+import { Config, queryTVL } from '../services'
 
 const BRIDGE_ADDRESS = '0xaBEA9132b05A70803a4E85094fD0e1800777fBEF'
